fix: debug settings restored after refresh. (#10)
Closes #8 

## Changes:
This PR introduces restoring debug settings across refreshes/hot-reload and open/close of the dev tools panel. 

As described in issue #5 I had to move the heartbeat event to the background script to make this work. In that process many of the features of the extension had to be moved as well but I believe that everything is working as intended.
I have also removed some dead code. Let me know if you would rather keep it.

Finally I added a changelog to reflect the changes since I intend to look at other items on the wishlist 🙂

## Reproduce steps
These are the steps I used to test it:
1. Remove “.parcel-cache” (the cache caused me some troubles). 
2. Build for Chrome or Firefix.
3. Load unpacked into the browser and load up an Excalibur project.
4. Open the dev panel, toggle debug and change the debug settings. 
5. Refresh or save a file to activate bundler (I used Vite). 
6. Debug settings should be still be activated after refresh.
7. The dev panel can also be closed and re-opened and the debug settings should reflect the current changes.